% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{JASPAR2022-class}
\alias{JASPAR2022-class}
\alias{JASPAR2022}
\title{JASPAR2022 object class}
\description{
The JASPAR2022 object class is a thin class for storing the path of
JASPAR2022 style SQLite file.
}
\section{Slots}{

\describe{
\item{\code{db}}{Object of class \code{"character"} a character string of the path
of SQLite file.}
}}

\examples{
library(TFBSTools)
opts <- list()
opts[["species"]] <- 9606
opts[["type"]] <- "SELEX"
opts[["all_versions"]] <- TRUE
PFMatrixList <- getMatrixSet(JASPAR2022, opts)

opts2 <- list()
opts2[["type"]] <- "SELEX"
PFMatrixList2 <- getMatrixSet(JASPAR2022, opts2)

}
\author{
Damir Baranasic
}
\keyword{classes}
